Move database handling files to their own directory.
