+++ Backend/app/database/initialize_db.py
@@ -1,7 +1,6 @@
 from pathlib import Path
-import sqlite3
 
-from app.database_actions import execute_in_database
+from app.database.database_actions import execute_in_database
 
 
 def initialize_database(db_path: Path) -> None:
